LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select/select-dialog-mode-focus.optional-expected.txt


Elephant
Tiger
Kangaroo
Dolphin
Giraffe
Panda
Zebra  Koala
Hippopotamus
Rhinoceros
Flamingo
Crocodile
Polar Bear
 Chimpanzee
 Ostrich
 Wolf

FAIL In dialog mode the first focusable element should get focus. assert_equals: The anchor should be focused. expected Element node <a id="interactive1" href="https://www.example.com/">Elep... but got Element node <select id="target">
  <div></div>
  <span></span>
  <a i...
FAIL In dialog mode tab should not close the picker. assert_false: The select should initially be closed. expected false got true
